refactor: move functions from torch
