+++ .pre-commit-config.yaml
@@ -7,7 +7,7 @@ repos:
     rev: 1.5.0
     hooks:
       - id: interrogate
-        args: ["--fail-under=50", "-v"]
+        args: ["--fail-under=80", "-v"]
 -   repo: https://github.com/PyCQA/pydocstyle
     rev: 6.1.1
     hooks:
