+++ proteinflow/data/torch.py
@@ -1,7 +1,6 @@
 import os
 import pickle
 import random
-import shutil
 from collections import defaultdict
 from copy import deepcopy
 from itertools import combinations
@@ -14,11 +13,6 @@ from tqdm import tqdm
 
 from proteinflow.constants import ALPHABET, CDR_REVERSE, D3TO1, MAIN_ATOMS
 from proteinflow.data import ProteinEntry
-from proteinflow.download.boto import (
-    _download_dataset_dicts_from_s3,
-    _download_dataset_from_s3,
-    _get_s3_paths_from_tag,
-)
 
 """Protein loader class inherited from `torch.utils.data.DataLoader`."""
 
@@ -903,174 +897,3 @@ class ProteinDataset(Dataset):
         if cdr is not None:
             data["cdr_id"] = CDR_REVERSE[cdr]
         return data
-
-
-def _download_dataset(tag, local_datasets_folder="./data/"):
-    """
-    Download the pre-processed data and the split dictionaries
-
-    Parameters
-    ----------
-    tag : str
-        name of the dataset (check `get_available_tags` to see the options)
-    local_dataset_folder : str, default "./data/"
-        the local folder that will contain proteinflow dataset folders, temporary files and logs
-
-    Returns
-    -------
-    data_folder : str
-        the path to the downloaded data folder
-    """
-
-    s3_data_path, s3_dict_path = _get_s3_paths_from_tag(tag)
-    data_folder = os.path.join(local_datasets_folder, f"proteinflow_{tag}")
-    dict_folder = os.path.join(
-        local_datasets_folder, f"proteinflow_{tag}", "splits_dict"
-    )
-
-    print("Downloading dictionaries for splitting the dataset...")
-    _download_dataset_dicts_from_s3(dict_folder, s3_dict_path)
-    print("Done!")
-
-    _download_dataset_from_s3(dataset_path=data_folder, s3_path=s3_data_path)
-    return data_folder
-
-
-def _biounits_in_clusters_dict(clusters_dict, excluded_files=None):
-    """
-    Return the list of all biounit files present in clusters_dict
-    """
-
-    if len(clusters_dict) == 0:
-        return np.array([])
-    if excluded_files is None:
-        excluded_files = []
-    return np.unique(
-        [
-            c[0]
-            for c in list(np.concatenate(list(clusters_dict.values())))
-            if c[0] not in excluded_files
-        ]
-    )
-
-
-def _exclude(clusters_dict, set_to_exclude, exclude_based_on_cdr=None):
-    """
-    Exclude biounits from clusters_dict
-
-    Parameters
-    ----------
-    clusters_dict : dict
-        dictionary of clusters
-    set_to_exclude : set
-        set of biounits to exclude
-    exclude_based_on_cdr : str, default None
-        if not None, exclude based only on clusters based on this CDR (e.g. "H3")
-
-    """
-    excluded_set = set()
-    excluded_dict = defaultdict(set)
-    for cluster in list(clusters_dict.keys()):
-        files = clusters_dict[cluster]
-        exclude = False
-        for biounit in files:
-            if biounit in set_to_exclude:
-                exclude = True
-                break
-        if exclude:
-            if exclude_based_on_cdr is not None:
-                if not cluster.endswith(exclude_based_on_cdr):
-                    continue
-            for biounit in files:
-                clusters_dict[cluster].remove(biounit)
-                excluded_dict[cluster].add(biounit)
-                excluded_set.add(biounit)
-    return excluded_dict, excluded_set
-
-
-def _split_data(
-    dataset_path="./data/proteinflow_20221110/",
-    excluded_files=None,
-    exclude_clusters=False,
-    exclude_based_on_cdr=None,
-):
-    """
-    Rearrange files into folders according to the dataset split dictionaries at `dataset_path/splits_dict`
-
-    Parameters
-    ----------
-    dataset_path : str, default "./data/proteinflow_20221110/"
-        The path to the dataset folder containing pre-processed entries and a `splits_dict` folder with split dictionaries (downloaded or generated with `get_split_dictionaries`)
-    excluded_files : list, optional
-        A list of files to exclude from the dataset
-    exclude_clusters : bool, default False
-        If True, exclude all files in a cluster if at least one file in the cluster is in `excluded_files`
-    exclude_based_on_cdr : str, optional
-        If not `None`, exclude all files in a cluster if the cluster name does not end with `exclude_based_on_cdr`
-    """
-
-    if excluded_files is None:
-        excluded_files = []
-
-    dict_folder = os.path.join(dataset_path, "splits_dict")
-    with open(os.path.join(dict_folder, "train.pickle"), "rb") as f:
-        train_clusters_dict = pickle.load(f)
-    with open(os.path.join(dict_folder, "valid.pickle"), "rb") as f:
-        valid_clusters_dict = pickle.load(f)
-    with open(os.path.join(dict_folder, "test.pickle"), "rb") as f:
-        test_clusters_dict = pickle.load(f)
-
-    train_biounits = _biounits_in_clusters_dict(train_clusters_dict, excluded_files)
-    valid_biounits = _biounits_in_clusters_dict(valid_clusters_dict, excluded_files)
-    test_biounits = _biounits_in_clusters_dict(test_clusters_dict, excluded_files)
-    train_path = os.path.join(dataset_path, "train")
-    valid_path = os.path.join(dataset_path, "valid")
-    test_path = os.path.join(dataset_path, "test")
-
-    if not os.path.exists(dataset_path):
-        os.makedirs(dataset_path)
-
-    if not os.path.exists(train_path):
-        os.makedirs(train_path)
-    if not os.path.exists(valid_path):
-        os.makedirs(valid_path)
-    if not os.path.exists(test_path):
-        os.makedirs(test_path)
-
-    if len(excluded_files) > 0:
-        set_to_exclude = set(excluded_files)
-        excluded_files = set()
-        excluded_clusters_dict = defaultdict(set)
-        if exclude_clusters:
-            for clusters_dict in [
-                train_clusters_dict,
-                valid_clusters_dict,
-                test_clusters_dict,
-            ]:
-                subset_excluded_set, subset_excluded_dict = _exclude(
-                    clusters_dict, set_to_exclude, exclude_based_on_cdr
-                )
-                excluded_files.update(subset_excluded_set)
-                excluded_clusters_dict.update(subset_excluded_dict)
-        excluded_files.update(set_to_exclude)
-        excluded_clusters_dict = {k: list(v) for k, v in excluded_clusters_dict.items()}
-        excluded_path = os.path.join(dataset_path, "excluded")
-        if not os.path.exists(excluded_path):
-            os.makedirs(excluded_path)
-        print("Updating the split dictionaries...")
-        with open(os.path.join(dict_folder, "train.pickle"), "wb") as f:
-            pickle.dump(train_clusters_dict, f)
-        with open(os.path.join(dict_folder, "excluded.pickle"), "wb") as f:
-            pickle.dump(excluded_clusters_dict, f)
-        print("Moving excluded files...")
-        for biounit in tqdm(excluded_files):
-            shutil.move(os.path.join(dataset_path, biounit), excluded_path)
-    print("Moving files in the train set...")
-    for biounit in tqdm(train_biounits):
-        shutil.move(os.path.join(dataset_path, biounit), train_path)
-    print("Moving files in the validation set...")
-    for biounit in tqdm(valid_biounits):
-        shutil.move(os.path.join(dataset_path, biounit), valid_path)
-    print("Moving files in the test set...")
-    for biounit in tqdm(test_biounits):
-        shutil.move(os.path.join(dataset_path, biounit), test_path)
